be/requirements.txt
# CompleteBytePOS Backend Requirements
# SQLite version (default, no additional driver needed)
# For MySQL: add PyMySQL>=1.1.0
# For PostgreSQL: add psycopg2-binary>=2.9.0

Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
python-barcode[images]>=0.15.0
qrcode[pil]>=7.4.0

# Development tools
ipython>=8.0.0
django-extensions>=3.2.0

# Production server
gunicorn>=21.2.0
whitenoise>=6.6.0  # For serving static files in production

